Add CleanWebpackPlugin and update webpack configuration
A CleanWebpackPlugin has been added to the plugins array for better manageability of the output folder during builds. In addition, the webpack configuration has been updated to include alias resolution for components, hooks, utils, styles, and context, and to modify the TerserPlugin's options for handling comments. This enhances code readability and maintainability.